help_desk/wsgi.py
"""
WSGI config for help_desk project.

It exposes the WSGI callable as a module-level variable named ``application``.

For more information on this file, see
https://docs.djangoproject.com/en/1.11/howto/deployment/wsgi/
"""

import os
from dotenv import load_dotenv

from django.core.wsgi import get_wsgi_application

load_dotenv( os.path.join(os.path.dirname(os.path.dirname(__file__)), '.env') ) 
os.environ.setdefault("DJANGO_SETTINGS_MODULE", "help_desk.settings")

if os.getenv('DJANGO_SETTINGS_MODULE'):
 os.environ['DJANGO_SETTINGS_MODULE'] = os.getenv('DJANGO_SETTINGS_MODULE')

application = get_wsgi_application()
