requirement.txt
Django==3.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.5.0
psycopg2==2.8.6
PyJWT==1.7.1
python-dotenv==0.19.0
redis==3.5.3
requests==2.26.0
